final commit

+++ app/Http/Controllers/ListingController.php
@@ -59,6 +59,10 @@ class ListingController extends Controller
 
     // Update listing
         public function update(Request $request, Listing $listing) {
+             //authorize edit only for user
+     if($listing->user_id != auth()->id()) {
+          abort(403, 'unauthorized action');
+     }
         $formFields = $request->validate([
            'title' => 'required',
            'company' => ['required'],
@@ -80,6 +84,10 @@ class ListingController extends Controller
 
     // Delete listing
     public function destroy(Listing $listing) {
+           //authorize edit only for user
+     if($listing->user_id != auth()->id()) {
+          abort(403, 'unauthorized action');
+     }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
     }
